requirements.txt
aiohttp
aiosqlite
google-genai
GoogleNews
greenlet
kiteconnect
openai
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-instrumentation-aiohttp-client
opentelemetry-sdk
pandas
pandas_ta
plotly
pydantic
pydantic-settings
python-dotenv
pytz
robin_stocks
schedule
sqlmodel
streamlit
structlog
yfinance